requirements.txt
python-dateutil==2.4.2
django==1.8
django-easy-pjax==1.2.0
docutils==0.12
